# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: 美宜佳门店助手 · 统一 API
  version: 0.1.0-m0
  description: |
    美宜佳门店助手统一后端 API。覆盖 12 个业务模块、80+ 端点。

    **权威设计文档**：[`docs/planning/unified-api-spec.md`](/docs/planning/unified-api-spec.md)
    **数据库 schema**：[`docs/planning/database-schema-reference.md`](/docs/planning/database-schema-reference.md)

    ## 全局约定（必读）

    - 所有接口都在 `/api/v1` 前缀下
    - **业务接口** 一律从 session 取 `storeId`（**不接受路径 / 参数里传**），spec § 0
    - **资源 CRUD 例外接口**（如 `PUT /master/stores/:id`）才显式带 `:id`
    - 认证靠 HttpOnly cookie `sso_token`，未登录受保护接口 → `401 UNAUTHORIZED`
    - session 没选门店时调业务接口 → `409 NO_STORE_SELECTED`（前端应跳门户首页让用户选店）

    ## 错误码

    | code | HTTP | 含义 |
    |---|---|---|
    | `UNAUTHORIZED` | 401 | 未登录 / token 失效 |
    | `FORBIDDEN` | 403 | 已登录但角色不够 |
    | `NOT_FOUND` | 404 | 资源不存在 |
    | `BAD_REQUEST` | 400 | 参数校验失败 |
    | `CONFLICT` | 409 | 业务冲突（如重复提交） |
    | `NO_STORE_SELECTED` | 409 | session 没有 currentStoreId |
    | `UPSTREAM_ERROR` | 502 | 调外部服务失败（飞书 / Dify / OpenRouter） |
    | `INTERNAL` | 500 | 兜底 |
  contact:
    name: rollingai-myj
    url: https://github.com/rollingai-myj/integratedApp
servers:
  - url: http://localhost:8787/api/v1
    description: 本地开发
  - url: /api/v1
    description: 当前部署
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sso_token
      description: HttpOnly cookie，登录成功后由后端自动种入
  responses:
    Unauthorized:
      description: 未登录 / token 失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorBody' }
          example:
            error: { code: UNAUTHORIZED, message: 请先登录 }
            requestId: req_abc123
    Forbidden:
      description: 已登录但角色不够
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorBody' }
...
    NoStoreSelected:
      description: session 里没选门店，前端应跳门户首页
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorBody' }
          example:
            error:
              code: NO_STORE_SELECTED
              message: 请先选择门店再操作（POST /portal/active-store）
  schemas:
...
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: BAD_REQUEST }
            message: { type: string, example: skuCode 必填 }
            details: { type: object, additionalProperties: true }
        requestId: { type: string, example: req_abc123 }
...
    StoreRef:
      type: object
      required: [id, code, name]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: 粤37893 }
        name: { type: string, example: 光明南店 }
        isPrimary: { type: boolean, example: true }
...
    PortalModulesResponse:
      type: object
      required: [modules]
      properties:
        modules:
          type: array
          items:
            type: object
            required: [key, label, enabled]
            properties:
              key: { type: string, enum: [shelves, prices, posters, admin] }
              label: { type: string }
              enabled: { type: boolean }
              disabledReason: { type: string, nullable: true }
    PortalStoresResponse:
      type: object
      required: [stores]
      properties:
        stores:
          type: array
          items: { $ref: '#/components/schemas/StoreRef' }
    SwitchStoreRequest:
      type: object
      required: [storeId]
      properties:
        storeId: { type: string, format: uuid }
    SwitchStoreResponse:
      type: object
      required: [currentStore]
      properties:
        currentStore: { $ref: '#/components/schemas/StoreRef' }

    # ── 主数据 ───────────────────────────────────────────────────────────
...
    PriceCurvePoint:
      type: object
      required: [snapshotDate, source]
      properties:
        snapshotDate: { type: string, format: date }
        retailPrice: { type: number, nullable: true }
        originalPrice: { type: number, nullable: true }
        wholesalePrice: { type: number, nullable: true }
        salesQty30d: { type: integer, nullable: true }
        salesAmount30d: { type: number, nullable: true }
        grossMargin30d: { type: number, nullable: true }
        source: { type: string, enum: [erp_sync, price_change, manual] }
        priceChangeId: { type: string, format: uuid, nullable: true }
    PriceCurveSku:
      type: object
      required: [skuCode, points]
      properties:
        skuCode: { type: string }
        productName: { type: string, nullable: true }
        points:
          type: array
          items: { $ref: '#/components/schemas/PriceCurvePoint' }
    PriceCurveResponse:
      type: object
      required: [curves]
      properties:
        curves:
          type: array
          items: { $ref: '#/components/schemas/PriceCurveSku' }
...
  parameters:
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 50 }
      description: 返回条数上限
security:
  - cookieAuth: []

# =============================================================================
# 业务路径
# =============================================================================
paths:
...
  /portal/modules:
    get:
      tags: [门户 (Portal)]
      summary: 当前用户可点亮的模块（含 disabledReason）
      responses:
        '200':
          description: 模块清单
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalModulesResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /portal/stores:
    get:
      tags: [门户 (Portal)]
      summary: 当前用户可见门店列表
      description: 店主 = 绑定的门店；超管 = 全部门店
      responses:
        '200':
          description: 门店列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalStoresResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /portal/active-store:
    post:
      tags: [门户 (Portal)]
      summary: 切换当前激活门店
      description: |
        **唯一改 session.currentStoreId 的接口**。校验 storeId 在用户可见门店内（超管全过）。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SwitchStoreRequest' }
      responses:
        '200':
          description: 切店成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SwitchStoreResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ── 模块 3：门店主数据（资源 CRUD 例外区） ──────────────────────────────
...
  /prices/curve:
    get:
      tags: [价盘 (Prices)]
      summary: 本店指定 SKU 的价格 / 销量曲线
      parameters:
        - { name: skuCodes, in: query, schema: { type: string }, description: 逗号分隔，留空 = 全店所有 SKU }
        - { name: daysBack, in: query, schema: { type: integer, default: 90 } }
      responses:
        '200':
          description: 价格曲线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceCurveResponse' }
        '409': { $ref: '#/components/responses/NoStoreSelected' }
  /prices/changes:
    get:
      tags: [价盘 (Prices)]
      summary: 本店调价历史
      parameters:
        - { name: skuCode, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/LimitParam' }
      responses:
        '200': { description: OK }
        '409': { $ref: '#/components/responses/NoStoreSelected' }
...